Refactoring to file structure, and files
* Grouped stage classes into 'stage' package
* Moved exception classes into 'exceptions' package
* Made MatchType public
